业务删除增加验证

+++ carrental/src/main/java/com/carrental/sm/service/impl/system/BusinessService.java
@@ -18,6 +18,7 @@ import com.carrental.sm.common.PagerUtil;
 import com.carrental.sm.common.bean.Pager;
 import com.carrental.sm.dao.system.IBusinessDao;
 import com.carrental.sm.dao.system.ILogDao;
+import com.carrental.sm.dao.system.IRentCarDao;
 import com.carrental.sm.service.system.IBusinessService;
 
 /**
@@ -31,6 +32,8 @@ public class BusinessService implements IBusinessService {
 	private IBusinessDao businessDao;
 	@Autowired
 	private ILogDao logDao;
+	@Autowired
+	private IRentCarDao rentCarDao;
 
 	public List<Business> queryList(Business business, Pager pager) {
 		Map<String, Object> params = new HashMap<String, Object>();
@@ -76,7 +79,10 @@ public class BusinessService implements IBusinessService {
 	}
 
 	public String delete(String ids, String names, Admin loginUser) {
-		// TODO 验证是否被使用
+		// 验证是否被使用
+		if (this.rentCarDao.countByBusinessIds(ids) > 0) {
+			return "该业务已被关联使用，请不要删除";
+		}
 		this.businessDao.delete(ids);
 
 		Log log = new Log();
